main page and footer are added
